[REFACTOR] Extract classes into separate files
